Create FlopShellState struct
The FlopShellState struct holds information about the current
state of the shell, such as the PATH variable and the currently
mounted floppy image. This information can be passed to the command
execution function, so the command execution function will be able
to handle the execution of built-in and internal commands as well
as "normal" commands.